[FIRRTL] Split FullReset from InferResets; domain-scope MemToRegOfVec
Move FART out of firrtl-infer-resets into firrtl-full-reset so InferResets
only infers concrete reset types. FullReset builds domains, converts comb
mems in async full-reset domains only, then implements full reset.

Drop ConvertMemToRegOfVecAnnotation; remove MemToRegOfVec from the Firtool
pipeline. Keep firrtl-mem-to-reg-of-vec as a standalone utility for tests
and custom pipelines.
