--- test/Dialect/FIRRTL/SFCTests/directories.fir.memtoregofvec.anno.json
@@ -1,5 +1,0 @@
-[
-  {
-    "class": "sifive.enterprise.firrtl.ConvertMemToRegOfVecAnnotation$"
-  }
-]
