+++ test/Dialect/FIRRTL/SFCTests/mem-taps-reg.fir
@@ -3,10 +3,14 @@
 FIRRTL version 4.0.0
 circuit Top : %[[
   {
+    "class":"circt.FullResetAnnotation",
+    "target":"~Top|DUTModule>reset",
+    "resetType":"async"
+  },
+  {
     "class":"firrtl.transforms.DontTouchAnnotation",
     "target":"~Top|Top>memTap"
   },
-  {"class": "sifive.enterprise.firrtl.ConvertMemToRegOfVecAnnotation$"},
   {
     "class":"sifive.enterprise.grandcentral.MemTapAnnotation",
     "source":"~Top|DUTModule>rf",
@@ -24,7 +28,7 @@ circuit Top : %[[
 ]]
   module DUTModule :
     input clock : Clock
-    input reset : Reset
+    input reset : AsyncReset
     output io : { flip addr : UInt<3>, flip dataIn : UInt<8>, flip wen : UInt<1>, dataOut : UInt<8>}
 
     cmem rf : UInt<8> [8]
@@ -36,7 +40,7 @@ circuit Top : %[[
 
   public module Top :
     input clock : Clock
-    input reset : UInt<1>
+    input reset : AsyncReset
     output io : { flip addr : UInt<3>, flip dataIn : UInt<8>, flip wen : UInt<1>, dataOut : UInt<8>}
 
     inst dut of DUTModule
