+++ test/firtool/async-reset-anno.fir
@@ -1,7 +1,7 @@
-; RUN: firtool %s -parse-only | circt-opt -pass-pipeline='builtin.module(firrtl.circuit(firrtl-infer-resets))'  | FileCheck %s --check-prefixes COMMON,POST-INFER-RESETS
-; RUN: firtool %s -parse-only | circt-opt -pass-pipeline='builtin.module(firrtl.circuit(firrtl-infer-resets,firrtl.module(firrtl-sfc-compat)))' | FileCheck %s --check-prefixes COMMON,POST-SFC-COMPAT
+; RUN: firtool %s -parse-only | circt-opt -pass-pipeline='builtin.module(firrtl.circuit(firrtl-infer-resets,firrtl-full-reset))'  | FileCheck %s --check-prefixes COMMON,POST-FULL-RESET
+; RUN: firtool %s -parse-only | circt-opt -pass-pipeline='builtin.module(firrtl.circuit(firrtl-infer-resets,firrtl-full-reset,firrtl.module(firrtl-sfc-compat)))' | FileCheck %s --check-prefixes COMMON,POST-SFC-COMPAT
 
-; Check that FullResetAnnotation exists after infer-resets pass
+; Check that FullResetAnnotation exists after full-reset pass
 ; but is deleted after sfc-compat
 
 FIRRTL version 3.3.0
@@ -17,7 +17,7 @@ circuit test :%[[
   module test :
     input clock : Clock
     input reset : AsyncReset
-    ; POST-INFER-RESETS: [{class = "circt.FullResetAnnotation", resetType = "async"}]
+    ; POST-FULL-RESET: [{class = "circt.FullResetAnnotation", resetType = "async"}]
     ; POST-SFC-COMPAT-NOT: [{class = "circt.FullResetAnnotation", resetType = "async"}]
     input in : { foo : UInt<8>, bar : UInt<8>}
     output out : { foo : UInt<8>, bar : UInt<8>}
@@ -30,7 +30,7 @@ circuit test :%[[
     input in : { foo : UInt<8>, bar : UInt<8>}
     output out : { foo : UInt<8>, bar : UInt<8>}
 
-    ; POST-INFER-RESETS: [{class = "circt.FullResetAnnotation", resetType = "async"}]
+    ; POST-FULL-RESET: [{class = "circt.FullResetAnnotation", resetType = "async"}]
     ; POST-SFC-COMPAT-NOT: [{class = "circt.FullResetAnnotation", resetType = "async"}]
      wire r : AsyncReset
     connect out, in
