+++ test/firtool/memory.fir
@@ -7,9 +7,6 @@ FIRRTL version 4.0.0
 circuit Qux: %[[{
     "class": "sifive.enterprise.firrtl.MarkDUTAnnotation",
     "target":"~Qux|Qux"
-  },
-  {
-    "class": "sifive.enterprise.firrtl.ConvertMemToRegOfVecAnnotation$"
   }]]
   public module Qux:
     input clock: Clock
