+++ test/firtool/phase-ordering.fir
@@ -5,7 +5,7 @@
 
 ; CHECK-LABEL: firrtl.module @Issue794
 ; CHECK-SAME: (in %clock: !firrtl.clock,
-; CHECK:       %[[memory_0:.+]] = firrtl.reg %clock {{.*}}: !firrtl.clock, !firrtl.uint<8>
+; CHECK:       %[[memory_0:.+]] = firrtl.regreset %clock, %reset, {{.*}} : !firrtl.clock, !firrtl.asyncreset
 ; CHECK:       firrtl.mux(%[[v14:.+]], %wData_0, %[[memory_0]])
 ; CHECK:       firrtl.mux(%[[v19:.+]], %wData_1, %[[v5:.+]])
 ; CHECK:       firrtl.matchingconnect %[[memory_0]], %[[v22:.+]]
@@ -15,10 +15,13 @@ circuit Issue794: %[[{
     "target":"~Issue794|Issue794"
   },
   {
-    "class": "sifive.enterprise.firrtl.ConvertMemToRegOfVecAnnotation$"
+    "class":"circt.FullResetAnnotation",
+    "target":"~Issue794|Issue794>reset",
+    "resetType":"async"
   }]]
   public module Issue794:
     input clock: Clock
+    input reset: AsyncReset
     input rAddr: UInt<2>
     input rEn: UInt<1>
     output rData: UInt<8>
